% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNet.R
\name{bootNet}
\alias{bootNet}
\title{Essentially "bootnet", but with fitNetwork for all estimation}
\usage{
bootNet(
  data,
  m = NULL,
  nboots = 10,
  lags = NULL,
  caseDrop = FALSE,
  rule = "OR",
  ci = 0.95,
  caseMin = 0.05,
  caseMax = 0.75,
  caseN = 10,
  threshold = FALSE,
  fits = NULL,
  type = "g",
  saveMods = TRUE,
  verbose = TRUE,
  fitCoefs = FALSE,
  size = NULL,
  nCores = 1,
  cluster = "mclapply",
  block = FALSE,
  maxiter = 10,
  directedDiag = FALSE,
  beepno = NULL,
  dayno = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{m}{numeric}

\item{nboots}{numeric}

\item{lags}{numeric or logical}

\item{caseDrop}{logical}

\item{rule}{character}

\item{ci}{numeric}

\item{caseMin}{numeric}

\item{caseMax}{numeric}

\item{caseN}{numeric}

\item{threshold}{logical or numeric}

\item{fits}{logical?}

\item{type}{character}

\item{saveMods}{logical}

\item{verbose}{logical}

\item{fitCoefs}{logical}

\item{size}{numeric?}

\item{nCores}{numeric}

\item{cluster}{character}

\item{block}{logical}

\item{maxiter}{numeric}

\item{directedDiag}{logical}

\item{beepno}{something}

\item{dayno}{something}

\item{...}{other arguments}
}
\value{
A bootNet object
}
\description{
Just expanding tha bootnet
}
\examples{
1 + 1
}
