% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centAuto}
\alias{centAuto}
\title{Mimics centrality_auto}
\usage{
centAuto(x, which.net = "temporal", weighted = TRUE, signed = TRUE)
}
\arguments{
\item{x}{Not sure}

\item{which.net}{character}

\item{weighted}{logical}

\item{signed}{logical}
}
\value{
A list of stuff
}
\description{
Yep yep
}
\examples{
1 + 1
}
