% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralityClustering.R
\name{CentClust}
\alias{CentClust}
\alias{centAuto}
\alias{clustAuto}
\title{Node centrality and shortest path lengths}
\usage{
centAuto(x, which.net = "temporal", weighted = TRUE, signed = TRUE)

clustAuto(x, thresholdWS = 0, thresholdON = 0)
}
\arguments{
\item{x}{Output from one of the primary \code{modnets} functions. Can also
supply a list of network models, and the function will be applied to all
models in the list.}

\item{which.net}{Only applies to SUR networks, as well as those fit with the
\code{mlGVAR()} function. Character string to indicate which type of
network to compute centrality values for. Options are \code{"temporal"} for
the temporal network, \code{"contemporaneous"} for the contemporaneous
network, \code{"PDC"} for the partial directed correlation network, and
\code{"interactions"} for the temporal interaction network.}

\item{weighted}{Logical. If \code{TRUE} then results are converted to an
unweighted network.}

\item{signed}{Logical. Determines whether to ignore the signs of edges or
not. Primarily affects the output for expected influence statistics.}

\item{thresholdWS}{Numeric threshold for the WS values.}

\item{thresholdON}{Numeric threshold for the Zhang values.}
}
\value{
A list containing node centrality statistics, edge-betweenness
values, and shortest path lengths.
}
\description{
Mimics the \code{centrality_auto()} function from the \code{qgraph} package.
The purpose of amending this function was to make it compatible with outputs
from the \code{modnets} package. The main use of this function is as the
engine for the \code{centTable()} function.
}
\details{
Returns several node centrality statistics, edge-betweenness centrality, and
shortest path lengths. Betweenness and Closeness centrality are computed for
all types of networks, as well as edge-betweenness values and shortest path
lengths. For GGMs, Strength centrality and Expected Influence are also
computed. For SUR networks, InStrength, OutStrength, InExpectedInfluence, and
OutExpectedInfluence are computed instead.

The key distinction between this function and \code{centrality_auto()} is
that centrality values can be computed for the matrix of interactions within
a temporal network.
}
\examples{
\dontrun{
x <- fitNetwork(data)

centAuto(x)
}
}
