% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centPlot}
\alias{centPlot}
\title{Mimics centralityPlot from qgraph}
\usage{
centPlot(
  Wmats,
  scale = c("z-scores", "raw", "raw0", "relative"),
  which.net = "temporal",
  include = "all",
  labels = NULL,
  orderBy = NULL,
  decreasing = FALSE,
  plot = TRUE,
  verbose = TRUE,
  weighted = TRUE,
  signed = TRUE
)
}
\arguments{
\item{Wmats}{Models}

\item{scale}{character}

\item{which.net}{character}

\item{include}{character}

\item{labels}{character}

\item{orderBy}{character}

\item{decreasing}{logical}

\item{plot}{logical}

\item{verbose}{logical}

\item{weighted}{logical}

\item{signed}{logical}
}
\value{
Plot
}
\description{
Indeed
}
\examples{
1 + 1
}
