% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centTable}
\alias{centTable}
\title{Makes table of centrality values}
\usage{
centTable(
  Wmats,
  scale = TRUE,
  which.net = "temporal",
  labels = NULL,
  relative = FALSE,
  weighted = TRUE,
  signed = TRUE
)
}
\arguments{
\item{Wmats}{Model(s)}

\item{scale}{logical}

\item{which.net}{character}

\item{labels}{character}

\item{relative}{logical}

\item{weighted}{logical}

\item{signed}{logical}
}
\value{
Table
}
\description{
Lots of ways to do centrality
}
\examples{
1 + 1
}
