% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustAuto}
\alias{clustAuto}
\title{Mimics clustcoef_auto}
\usage{
clustAuto(x, thresholdWS = 0, thresholdON = 0)
}
\arguments{
\item{x}{something}

\item{thresholdWS}{something}

\item{thresholdON}{something}
}
\value{
A list of stuff
}
\description{
Mimics clustcoef_auto
}
\examples{
1 + 1
}
