% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustPlot}
\alias{clustPlot}
\title{Mimics clusteringPlot}
\usage{
clustPlot(
  Wmats,
  scale = c("z-scores", "raw", "raw0", "relative"),
  include = "all",
  labels = NULL,
  orderBy = NULL,
  decreasing = FALSE,
  plot = TRUE,
  signed = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Wmats}{Models}

\item{scale}{character}

\item{include}{character}

\item{labels}{character}

\item{orderBy}{character}

\item{decreasing}{logical}

\item{plot}{logical}

\item{signed}{logical}

\item{verbose}{logical}
}
\value{
Plot
}
\description{
Mimics clusteringPlot
}
\examples{
1 + 1
}
