% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustTable}
\alias{clustTable}
\title{Mimics clusteringTable from qgraph}
\usage{
clustTable(Wmats, scale = TRUE, labels = NULL, relative = FALSE, signed = TRUE)
}
\arguments{
\item{Wmats}{Models}

\item{scale}{logical}

\item{labels}{character}

\item{relative}{logical}

\item{signed}{logical}
}
\value{
Table
}
\description{
But with moderators!
}
\examples{
1 + 1
}
