% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerVAR.R
\name{compareVAR}
\alias{compareVAR}
\title{Compare two to three lmerVAR models}
\usage{
compareVAR(m1, m2, m3 = NULL, anova = NULL, type = "tempMods")
}
\arguments{
\item{m1}{model 1}

\item{m2}{model 2}

\item{m3}{model 3 (optional)}

\item{anova}{numeric?}

\item{type}{character}
}
\value{
Table of ANOVA
}
\description{
Gotta see how this works
}
\examples{
1 + 1
}
