% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{condPlot}
\alias{condPlot}
\title{Plot conditional effects}
\usage{
condPlot(
  out,
  to,
  from,
  swap = FALSE,
  avg = FALSE,
  compare = NULL,
  hist = FALSE,
  xlab = NULL,
  mods = NULL,
  nsims = 500,
  xn = NULL,
  getCIs = FALSE,
  discrete = FALSE,
  ylab = NULL,
  main = NULL,
  midline = TRUE
)
}
\arguments{
\item{out}{network}

\item{to}{y}

\item{from}{x}

\item{swap}{logical}

\item{avg}{logical}

\item{compare}{something}

\item{hist}{something}

\item{xlab}{character}

\item{mods}{something}

\item{nsims}{numeric}

\item{xn}{something}

\item{getCIs}{logical}

\item{discrete}{logical}

\item{ylab}{character}

\item{main}{character}

\item{midline}{logical}
}
\value{
Plot
}
\description{
Description
}
\examples{
1 + 1
}
