% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetwork.R
\name{covNet}
\alias{covNet}
\title{Weird function that I'm not sure will be necessary}
\usage{
covNet(object, mnet = TRUE, threshold = 0.05)
}
\arguments{
\item{object}{fitNetwork object}

\item{mnet}{logical}

\item{threshold}{numeric or logical}
}
\value{
A network
}
\description{
Seems to mostly be for plotting with plotCovNet
}
\examples{
1 + 1
}
