% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNet.R
\name{cscoef}
\alias{cscoef}
\title{Return correlation stability coefficients from caseDrop procedure}
\usage{
cscoef(obj, cor = 0.7, ci = 0.95, first = TRUE, verbose = TRUE)
}
\arguments{
\item{obj}{bootNet object}

\item{cor}{numeric}

\item{ci}{numeric}

\item{first}{logical}

\item{verbose}{logical}
}
\value{
CS coefficients
}
\description{
Return correlation stability coefficients from caseDrop procedure
}
\examples{
1 + 1
}
