% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetwork.R
\name{fitNetwork}
\alias{fitNetwork}
\title{Fit cross-sectional and idiographic moderated network models}
\usage{
fitNetwork(
  data,
  moderators = NULL,
  type = "gaussian",
  lags = NULL,
  seed = NULL,
  folds = 10,
  gamma = 0.5,
  which.lam = "lambda.min",
  rule = "OR",
  threshold = FALSE,
  scale = FALSE,
  std = TRUE,
  center = TRUE,
  covariates = NULL,
  verbose = FALSE,
  exogenous = TRUE,
  mval = NULL,
  residMat = "sigma",
  medges = 1,
  pcor = FALSE,
  maxiter = 100,
  getLL = TRUE,
  saveMods = TRUE,
  binarize = FALSE,
  fitCoefs = FALSE,
  detrend = FALSE,
  beepno = NULL,
  dayno = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{moderators}{numeric}

\item{type}{character}

\item{lags}{numeric or logical}

\item{seed}{numeric}

\item{folds}{numeric}

\item{gamma}{numeric}

\item{which.lam}{character}

\item{rule}{character}

\item{threshold}{numeric or logical}

\item{scale}{logical}

\item{std}{logical}

\item{center}{logical}

\item{covariates}{numeric or list}

\item{verbose}{logical}

\item{exogenous}{logical}

\item{mval}{numeric}

\item{residMat}{character}

\item{medges}{numeric}

\item{pcor}{logical}

\item{maxiter}{numeric}

\item{getLL}{logical}

\item{saveMods}{logical}

\item{binarize}{logical}

\item{fitCoefs}{logical}

\item{detrend}{logical}

\item{beepno}{something}

\item{dayno}{something}

\item{...}{other arguments}
}
\value{
A ggm or SURnetwork
}
\description{
The main function that ties everything together for both cross-sectional and
idiographic (temporal) network models, moderated or otherwise.
}
\examples{
1 + 1
}
