% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggmDat}
\alias{ggmDat}
\title{Simulated GGM data}
\format{
A \code{5000 x 6} data frame with five variables as nodes in a GGM
and a sixth as a moderator. Attributes contain values for the data
generating model. \code{b1} contains the pairwise network, while \code{b2}
contains the interaction matrix. \code{intercepts} contain the population
means of the five node variables. \code{m} contains the population mean of
the moderator. \code{m1} contains the main effect coefficients for the
moderator predicting each of the five nodes.
}
\usage{
ggmDat
}
\description{
Data generated from \code{\link{simNet}}, with five variables that serve as
nodes in the GGM and a sixth that serves as a moderator.
}
\keyword{datasets}
