% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvarDat}
\alias{gvarDat}
\title{Simulated temporal network data}
\format{
A \code{5000 x 6} data frame with five variables as nodes in a GVAR
and a sixth as a moderator. Attributes contain values for the data
generating model.
}
\usage{
gvarDat
}
\description{
Data generated from \code{\link{simNet}}, with five variables that serve as
nodes in the GVAR model and a sixth that serves as a moderator. The data were
generated by setting \code{lag = 1} to specify a single-subject lag-1 model.
The data were simulated from a graphical vector autoregressive model (GVAR).
}
\details{
In the case of using \code{simNet} with \code{lags = 1}, that function
essentially serves as a wrapper for \code{\link{mlGVARsim}} which
automatically sets \code{nPerson = 1}.
}
\keyword{datasets}
