% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{intsPlot}
\alias{intsPlot}
\title{Plot CI difference for each predictor with interactions against Y}
\usage{
intsPlot(out, y = "all", moderator = NULL, nsims = 500, alpha = 0.05)
}
\arguments{
\item{out}{network}

\item{y}{outcome, character}

\item{moderator}{something}

\item{nsims}{numeric}

\item{alpha}{numeric}
}
\value{
Plot
}
\description{
This shows something about all the interaction terms
}
\examples{
1 + 1
}
