% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerVAR.R
\name{lmerNets}
\alias{lmerNets}
\title{Create networks out of lmerVAR models}
\usage{
lmerNets(model, inds, m = NULL, threshold = FALSE, rule = "OR", ggm = "pcor")
}
\arguments{
\item{model}{output from lmerVAR?}

\item{inds}{something}

\item{m}{mnames}

\item{threshold}{logical numeric}

\item{rule}{character}

\item{ggm}{character}
}
\value{
lmerNetworks
}
\description{
Gotta look into how this works
}
\examples{
1 + 1
}
