% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerVAR.R
\name{lmerVAR}
\alias{lmerVAR}
\title{Legit mixed-effects modeling for the GVAR in multilevel data}
\usage{
lmerVAR(
  data,
  m = NULL,
  temporal = "default",
  contemp = "default",
  idvar = "ID",
  intvars = NULL,
  center = TRUE,
  scale = TRUE,
  centerWithin = TRUE,
  scaleWithin = FALSE,
  exogenous = TRUE,
  covariates = NULL,
  fix = NULL,
  warnings = FALSE,
  verbose = TRUE,
  beepno = NULL,
  dayno = NULL,
  deleteMissing = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{m}{numeric}

\item{temporal}{character}

\item{contemp}{character}

\item{idvar}{character}

\item{intvars}{character vector}

\item{center}{logical}

\item{scale}{logical}

\item{centerWithin}{logical}

\item{scaleWithin}{logical}

\item{exogenous}{logical}

\item{covariates}{character or list}

\item{fix}{something}

\item{warnings}{logical}

\item{verbose}{logical}

\item{beepno}{something}

\item{dayno}{something}

\item{deleteMissing}{logical}
}
\value{
lmerVAR models
}
\description{
Something or other
}
\examples{
1 + 1
}
