% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlGVAR.R
\name{mlGVAR}
\alias{mlGVAR}
\title{Fit GVAR models with multilevel data}
\usage{
mlGVAR(
  data,
  m = NULL,
  selectFUN = NULL,
  subjectNets = FALSE,
  idvar = "ID",
  exogenous = TRUE,
  center = TRUE,
  scale = TRUE,
  fixedType = "g",
  betweenType = "g",
  centerWithin = TRUE,
  scaleWithin = FALSE,
  rule = "OR",
  threshold = "none",
  verbose = TRUE,
  pcor = FALSE,
  fixedArgs = NULL,
  betweenArgs = NULL,
  bm = FALSE,
  beepno = NULL,
  dayno = NULL,
  deleteMissing = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{m}{numeric}

\item{selectFUN}{character?}

\item{subjectNets}{logical}

\item{idvar}{character}

\item{exogenous}{logical}

\item{center}{logical}

\item{scale}{logical}

\item{fixedType}{character}

\item{betweenType}{character}

\item{centerWithin}{logical}

\item{scaleWithin}{logical}

\item{rule}{character}

\item{threshold}{character or numeric or logical}

\item{verbose}{logical}

\item{pcor}{logical}

\item{fixedArgs}{list}

\item{betweenArgs}{list}

\item{bm}{logical}

\item{beepno}{something}

\item{dayno}{something}

\item{deleteMissing}{logical}

\item{...}{other arguments}
}
\value{
ML-GVAR models
}
\description{
Pseudo mixed-effects model. No real estimation of random effects.
}
\examples{
1 + 1
}
