% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlgvarDat}
\alias{mlgvarDat}
\title{Simulated multi-level network data}
\format{
A \code{50000 x 7} data frame with five variables to serve as nodes in
the networks, one variable \code{"M"} to serve as the time-lagged
moderator, and an \code{ID} variable that labels responses from each of
\code{100} simulated individuals. For each ID number, there are 500
time-ordered responses.
}
\usage{
mlgvarDat
}
\description{
Data generated from \code{\link{mlGVARsim}}, with five variables that serve
as nodes in the multi-level GVAR model, one moderator variable, and an ID
variable that distinguishes between subjects.
}
\keyword{datasets}
