% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlGVAR.R
\name{mlGVARsim}
\alias{mlGVARsim}
\title{Main workhorse for simulating VAR and mlGVAR data}
\usage{
mlGVARsim(
  nTime = 50,
  nPerson = 10,
  nNode = 3,
  m = NULL,
  m2 = 0.25,
  m1 = 0.7,
  m0 = 1,
  lag = 1,
  thetaVar = NULL,
  mu_SD = NULL,
  init_beta_SD = NULL,
  fixedMuSD = 1,
  shrink_fixed = 0.9,
  propPos = 0.5,
  m1SD = 0.1,
  m2SD = 0.1,
  m1_range = NULL,
  m2_range = NULL,
  shrink_deviation = 0.9,
  getM = FALSE,
  contemporaneous = "wishart",
  GGMsparsity = 0.5,
  mcenter = TRUE,
  skew = FALSE,
  skewErr = FALSE,
  ordinal = FALSE,
  nLevels = 5,
  ordWithin = TRUE,
  minOrd = 3,
  thresholds = NULL,
  mseed = NULL,
  onlyNets = FALSE,
  modType = "none"
)
}
\arguments{
\item{nTime}{numeric}

\item{nPerson}{numeric}

\item{nNode}{numeric}

\item{m}{numeric or logical}

\item{m2}{numeric}

\item{m1}{numeric}

\item{m0}{numeric}

\item{lag}{numeric or logical}

\item{thetaVar}{numeric}

\item{mu_SD}{numeric}

\item{init_beta_SD}{numeric}

\item{fixedMuSD}{numeric}

\item{shrink_fixed}{numeric}

\item{propPos}{numeric}

\item{m1SD}{numeric}

\item{m2SD}{numeric}

\item{m1_range}{numeric, two values}

\item{m2_range}{numeric, two values}

\item{shrink_deviation}{numeric}

\item{getM}{logical}

\item{contemporaneous}{character}

\item{GGMsparsity}{numeric}

\item{mcenter}{logical}

\item{skew}{logical}

\item{skewErr}{logical}

\item{ordinal}{logical}

\item{nLevels}{numeric}

\item{ordWithin}{logical}

\item{minOrd}{numeric}

\item{thresholds}{numeric vector}

\item{mseed}{numeric}

\item{onlyNets}{logical}

\item{modType}{character}
}
\value{
Simulated mlGVAR or VAR data
}
\description{
Description
}
\examples{
1 + 1
}
