% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnetPowerSim.R
\name{mnetPowerSim}
\alias{mnetPowerSim}
\title{Power simulator for cross-sectional and idiographic networks}
\usage{
mnetPowerSim(
  niter = 10,
  N = 100,
  p = 5,
  m = FALSE,
  m1 = 0,
  m2 = 0.1,
  sparsity = 0.5,
  lags = NULL,
  trueNet = NULL,
  threshold = TRUE,
  rule = "OR",
  avg = TRUE,
  maxiter = 100,
  saveFits = TRUE,
  saveData = FALSE,
  intercepts = NULL,
  mbinary = FALSE,
  select = NULL,
  vargs = list(),
  type = "g",
  gibbs = TRUE,
  ordinal = FALSE,
  mord = FALSE,
  nLevels = 5,
  minOrd = 3,
  div = 1000,
  modType = "none",
  m1_range = NULL,
  m2_range = c(0.1, 0.3),
  time = TRUE,
  skewErr = FALSE,
  nCores = 1,
  cluster = "mclapply",
  fixedPar = NULL,
  V2 = 1,
  ...
)
}
\arguments{
\item{niter}{numeric}

\item{N}{numeric}

\item{p}{numeric}

\item{m}{logical or numeric?}

\item{m1}{numeric}

\item{m2}{numeric}

\item{sparsity}{numeric}

\item{lags}{numeric}

\item{trueNet}{matrix? list?}

\item{threshold}{numeric}

\item{rule}{character}

\item{avg}{logical}

\item{maxiter}{numeric}

\item{saveFits}{logical}

\item{saveData}{logical}

\item{intercepts}{numeric vector}

\item{mbinary}{logical}

\item{select}{something}

\item{vargs}{list}

\item{type}{character}

\item{gibbs}{logical}

\item{ordinal}{logical}

\item{mord}{logical}

\item{nLevels}{numeric}

\item{minOrd}{numeric}

\item{div}{numeric}

\item{modType}{character}

\item{m1_range}{numeric}

\item{m2_range}{numeric}

\item{time}{logical}

\item{skewErr}{logical}

\item{nCores}{numeric}

\item{cluster}{character}

\item{fixedPar}{something}

\item{V2}{numeric}

\item{...}{Additional arguments.}
}
\value{
Power simulation results
}
\description{
Simulation process for power
}
\examples{
1 + 1
}
