% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLL.R
\name{modLL}
\alias{modLL}
\title{Omnibus log-likelihood of a network and LRT comparing two networks.}
\usage{
modLL(
  net0,
  net1 = NULL,
  nodes = FALSE,
  lrt = NULL,
  all = FALSE,
  d = 4,
  alpha = 0.05,
  orderBy = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{net0}{GGM}

\item{net1}{GGM or null}

\item{nodes}{logical}

\item{lrt}{logical}

\item{all}{logical}

\item{d}{numeric}

\item{alpha}{numeric}

\item{orderBy}{character}

\item{decreasing}{logical}
}
\value{
A named vector
}
\description{
Computes log-likelihood, AIC, and BIC for a whole network.
Also compares two networks using a likelihoo ratio test.
}
\examples{
1 + 1
}
