% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modnets.R
\docType{package}
\name{modnets}
\alias{modnets}
\title{modnets: A package for fitting moderated network models}
\description{
The modnets package is all about fitting and analyzing moderated network
models. Currently in development -- more updates and details to come.
}
