% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{modSelect}
\alias{modSelect}
\title{Select a model from the resample function}
\usage{
modSelect(
  obj,
  data = NULL,
  fit = FALSE,
  select = "select",
  thresh = NULL,
  ascall = TRUE,
  type = "gaussian",
  ...
)
}
\arguments{
\item{obj}{resample object}

\item{data}{data}

\item{fit}{logical}

\item{select}{character}

\item{thresh}{numeric}

\item{ascall}{logical}

\item{type}{character}

\item{...}{Additional arguments.}
}
\value{
a list of information for fitNetwork
}
\description{
Creates the necessary input for fitNetwork when selecting variables.
}
\examples{
1 + 1
}
