% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLL.R
\name{modTable}
\alias{modTable}
\title{Moderated network table}
\usage{
modTable(
  fits,
  nodes = FALSE,
  orderBy = TRUE,
  d = 4,
  alpha = 0.05,
  decreasing = TRUE,
  names = NULL,
  rmsea = FALSE
)
}
\arguments{
\item{fits}{list of network models}

\item{nodes}{logical}

\item{orderBy}{character}

\item{d}{numeric}

\item{alpha}{numeric}

\item{decreasing}{logical}

\item{names}{logical}

\item{rmsea}{logical}
}
\value{
A list of tables
}
\description{
Fits LRT to a list of network models to compare them all against each other.
}
\examples{
1 + 1
}
