% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net}
\alias{net}
\alias{netInts}
\title{Get adjacency matrices from fit objects}
\usage{
net(
  fit,
  n = "beta",
  threshold = FALSE,
  rule = "OR",
  binary = FALSE,
  nodewise = FALSE,
  d = 14,
  r = NULL
)

netInts(
  fit,
  n = "temporal",
  threshold = FALSE,
  avg = FALSE,
  rule = "none",
  r = NULL,
  empty = TRUE,
  mselect = NULL
)
}
\arguments{
\item{fit}{model}

\item{n}{character}

\item{threshold}{numeric or logical}

\item{rule}{chatacter}

\item{binary}{logical}

\item{nodewise}{logical}

\item{d}{numeric}

\item{r}{something}

\item{avg}{logical}

\item{empty}{logical}

\item{mselect}{logical}
}
\value{
A network matrix
}
\description{
Used all the time
}
\examples{
1 + 1
}
