% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordinalize}
\alias{ordinalize}
\title{Turn continuous data into ordinal data}
\usage{
ordinalize(
  data,
  m = NULL,
  nLevels = 5,
  thresholds = NULL,
  mthresh = NULL,
  mord = TRUE,
  minOrd = 3
)
}
\arguments{
\item{data}{data}

\item{m}{moderator}

\item{nLevels}{numeric}

\item{thresholds}{numeric}

\item{mthresh}{numeric}

\item{mord}{logical}

\item{minOrd}{numeric}
}
\value{
ordinal data
}
\description{
Yeah
}
\examples{
1 + 1
}
