% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R, R/plotMethods.R
\name{plotBoot}
\alias{plotBoot}
\alias{plot.bootNet}
\title{Plot results of bootNet}
\usage{
plotBoot(
  x,
  type = "edges",
  net = "temporal",
  plot = "all",
  cor = 0.7,
  order = "mean",
  ci = 0.95,
  pairwise = TRUE,
  interactions = TRUE,
  labels = NULL,
  title = NULL,
  cis = "quantile",
  true = NULL,
  errbars = FALSE,
  vline = FALSE,
  threshold = FALSE,
  difference = FALSE,
  color = FALSE,
  text = FALSE,
  textPos = "value",
  multi = NULL,
  directedDiag = FALSE,
  ...
)

\method{plot}{bootNet}(
  x,
  type = "edges",
  net = "temporal",
  plot = "all",
  cor = 0.7,
  order = "mean",
  ci = 0.95,
  pairwise = TRUE,
  interactions = TRUE,
  labels = NULL,
  title = NULL,
  cis = "quantile",
  true = NULL,
  errbars = FALSE,
  vline = FALSE,
  threshold = FALSE,
  difference = FALSE,
  color = FALSE,
  text = FALSE,
  textPos = "value",
  multi = NULL,
  directedDiag = FALSE,
  ...
)
}
\arguments{
\item{x}{network}

\item{type}{character}

\item{net}{character}

\item{plot}{character}

\item{cor}{numeric}

\item{order}{character}

\item{ci}{numeric}

\item{pairwise}{logical}

\item{interactions}{logical}

\item{labels}{character vector}

\item{title}{character}

\item{cis}{character}

\item{true}{logical}

\item{errbars}{logical}

\item{vline}{numeric or logical}

\item{threshold}{numeric or logical}

\item{difference}{logical}

\item{color}{logical}

\item{text}{logical}

\item{textPos}{character}

\item{multi}{logical}

\item{directedDiag}{logical}

\item{...}{other arguments}
}
\value{
A plot
}
\description{
Currently only available for caseDrop
}
\examples{
1 + 1
}
