% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCentrality}
\alias{plotCentrality}
\title{Plot centrality and clustering values for multiple networks}
\usage{
plotCentrality(
  Wmats,
  which.net = "temporal",
  scale = TRUE,
  labels = NULL,
  plot = TRUE,
  centrality = "all",
  clustering = "Zhang"
)
}
\arguments{
\item{Wmats}{Models}

\item{which.net}{character}

\item{scale}{logical}

\item{labels}{character}

\item{plot}{logical}

\item{centrality}{character or character vector}

\item{clustering}{character or character vector}
}
\value{
Plot
}
\description{
Can handle all kinds of (moderated) networks.
}
\examples{
1 + 1
}
