% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCoefs}
\alias{plotCoefs}
\title{Plot coefficients from 'SURfit' with confidence intervals}
\usage{
plotCoefs(
  fit,
  true = FALSE,
  alpha = 0.05,
  plot = TRUE,
  col = "blue",
  flip = TRUE,
  data = NULL,
  select = TRUE,
  size = 1,
  labels = TRUE,
  title = NULL,
  vars = "all"
)
}
\arguments{
\item{fit}{SURfit object}

\item{true}{logical}

\item{alpha}{numeric}

\item{plot}{logical}

\item{col}{character}

\item{flip}{logical}

\item{data}{data}

\item{select}{logical}

\item{size}{numeric}

\item{labels}{logical}

\item{title}{character}

\item{vars}{character or numeric}
}
\value{
A plot
}
\description{
Plotting. I think this actually goes with resample? Also may not be restricted
to SURfit objects
}
\examples{
1 + 1
}
