% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCovNet}
\alias{plotCovNet}
\title{Plot network using output from covNet}
\usage{
plotCovNet(object, border = NULL, color = NULL, ...)
}
\arguments{
\item{object}{network}

\item{border}{something}

\item{color}{character}

\item{...}{other arguments}
}
\value{
Plot
}
\description{
Description
}
\examples{
1 + 1
}
