% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotMods}
\alias{plotMods}
\title{Better way to visualize mvals}
\usage{
plotMods(
  nets,
  nodewise = FALSE,
  elsize = 2,
  vsize = NULL,
  elabs = TRUE,
  predict = NULL,
  layout = NULL,
  which.net = "temporal",
  ...
)
}
\arguments{
\item{nets}{list of networks}

\item{nodewise}{logical}

\item{elsize}{numeric}

\item{vsize}{numeric}

\item{elabs}{logical}

\item{predict}{something}

\item{layout}{character}

\item{which.net}{character}

\item{...}{other arguments}
}
\value{
Plot
}
\description{
Gotta check on this but seems cool
}
\examples{
1 + 1
}
