% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R, R/plotMethods.R
\name{plotNet}
\alias{plotNet}
\alias{plot.ggm}
\alias{plot.SURnet}
\alias{plot.mlGVAR}
\alias{plot.lmerVAR}
\alias{plot.mgmSim}
\alias{plot.mlVARsim}
\alias{plot.simMLgvar}
\title{Plot moderated network models}
\usage{
plotNet(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{ggm}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{SURnet}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{mlGVAR}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{lmerVAR}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{mgmSim}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{mlVARsim}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)

\method{plot}{simMLgvar}(
  x,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  covNet = FALSE,
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)
}
\arguments{
\item{x}{network}

\item{which.net}{character}

\item{threshold}{numeric or logical}

\item{layout}{character}

\item{predict}{logical or network}

\item{mnet}{logical}

\item{names}{logical}

\item{nodewise}{logical}

\item{scale}{logical}

\item{lag}{numeric or logical}

\item{con}{character}

\item{cat}{character}

\item{covNet}{logical}

\item{plot}{logical}

\item{elabs}{logical}

\item{elsize}{numeric}

\item{rule}{character}

\item{binarize}{logical}

\item{mlty}{logical}

\item{mselect}{something}

\item{...}{Additional arguments.}
}
\value{
A plot
}
\description{
Core function for plotting various types of network models.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: For ggms!
}}

\examples{
1 + 1
}
