% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotNet2}
\alias{plotNet2}
\title{Plot the temporal and contemporaneous networks in the same window}
\usage{
plotNet2(
  object,
  whichNets = NULL,
  whichTemp = c("temporal", "PDC"),
  titles = c("PDC ", "PCC "),
  ...
)
}
\arguments{
\item{object}{SUR output}

\item{whichNets}{Vector of which networks to plot}

\item{whichTemp}{Which version of the temporal network should be plotted}

\item{titles}{Titles}

\item{...}{other arguments}
}
\value{
A plot
}
\description{
Description
}
\examples{
1 + 1
}
