% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotNet3}
\alias{plotNet3}
\title{Easy plotting for mlGVAR and lmerVAR output}
\usage{
plotNet3(
  object,
  ...,
  nets = c("temporal", "contemporaneous", "between"),
  titles = TRUE,
  l = 3,
  label = NULL,
  xpos = 0,
  ypos = 0.5
)
}
\arguments{
\item{object}{mlGVAR or lmerVAR object}

\item{...}{other arguments}

\item{nets}{which networks to plot}

\item{titles}{character}

\item{l}{numeric}

\item{label}{something}

\item{xpos}{numeric}

\item{ypos}{numeric}
}
\value{
Plot
}
\description{
Description
}
\examples{
1 + 1
}
