% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R, R/plotMethods.R
\name{plotPower}
\alias{plotPower}
\alias{plot.mnetPower}
\title{Plot results of power simulation}
\usage{
plotPower(
  x,
  by = "type",
  yvar = "default",
  yadd = NULL,
  hline = 0.8,
  xlab = "Number of cases",
  title = NULL,
  ...
)

\method{plot}{mnetPower}(
  x,
  by = "type",
  yvar = "default",
  yadd = NULL,
  hline = 0.8,
  xlab = "Number of cases",
  title = NULL,
  ...
)
}
\arguments{
\item{x}{mnetPowerSim output}

\item{by}{character}

\item{yvar}{character ?}

\item{yadd}{something}

\item{hline}{numeric}

\item{xlab}{character}

\item{title}{character}

\item{...}{other arguments}
}
\value{
Plot
}
\description{
Good with mnetPowerSim
}
\examples{
1 + 1
}
