% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPvals}
\alias{plotPvals}
\title{Plot the ECDF of pvalues from resample}
\usage{
plotPvals(obj, outcome = 1, predictor = 1, alpha = 0.05)
}
\arguments{
\item{obj}{output from resample}

\item{outcome}{numeric}

\item{predictor}{numeric}

\item{alpha}{numeric}
}
\value{
A plot
}
\description{
Plots pvalues or something
}
\examples{
1 + 1
}
