% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotStability}
\alias{plotStability}
\title{Plot stability paths for a given variable}
\usage{
plotStability(obj, pp = 1, s = 3, thresh = 0.5, color = "black")
}
\arguments{
\item{obj}{output from resample}

\item{pp}{numeric}

\item{s}{numeric}

\item{thresh}{numeric}

\item{color}{character}
}
\value{
A plot
}
\description{
You heard me
}
\examples{
1 + 1
}
