% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNet.R
\name{predictNet}
\alias{predictNet}
\title{Calculate prediction error from network models}
\usage{
predictNet(object, data = NULL, all = FALSE, scale = FALSE)
}
\arguments{
\item{object}{network}

\item{data}{dataset, or network for comparison}

\item{all}{logical}

\item{scale}{logical}
}
\value{
Table
}
\description{
See the prediction error based on different statistics for either GGMs or SURs.
Also can compare and find the change values (such as R-squared change) between two networks
}
\examples{
1 + 1
}
