% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Bootstrapping or multi-sample splits for variable selection}
\usage{
resample(
  data,
  m = NULL,
  niter = 10,
  sampMethod = "bootstrap",
  criterion = "AIC",
  method = "glmnet",
  rule = "OR",
  gamma = 0.5,
  nfolds = 10,
  nlam = 50,
  which.lam = "min",
  threshold = FALSE,
  bonf = FALSE,
  alpha = 0.05,
  exogenous = TRUE,
  split = 0.5,
  center = TRUE,
  scale = FALSE,
  varSeed = NULL,
  seed = NULL,
  verbose = TRUE,
  lags = NULL,
  binary = NULL,
  type = "g",
  saveMods = TRUE,
  saveData = FALSE,
  saveVars = FALSE,
  fitit = TRUE,
  nCores = 1,
  cluster = "mclapply",
  block = FALSE,
  beepno = NULL,
  dayno = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{m}{numeric}

\item{niter}{numeric}

\item{sampMethod}{character}

\item{criterion}{character}

\item{method}{character}

\item{rule}{character}

\item{gamma}{numeric}

\item{nfolds}{numeric}

\item{nlam}{numeric}

\item{which.lam}{character}

\item{threshold}{numeric or logical}

\item{bonf}{logical}

\item{alpha}{numeric}

\item{exogenous}{logical}

\item{split}{numeric}

\item{center}{logical}

\item{scale}{logical}

\item{varSeed}{numeric}

\item{seed}{numeric}

\item{verbose}{logical}

\item{lags}{numeric or logical}

\item{binary}{logical}

\item{type}{character}

\item{saveMods}{logical}

\item{saveData}{logical}

\item{saveVars}{logical}

\item{fitit}{logical}

\item{nCores}{numeric}

\item{cluster}{character}

\item{block}{logical}

\item{beepno}{something}

\item{dayno}{something}

\item{...}{other arguments}
}
\value{
Lots of stuff
}
\description{
Multiple resampling procedures for selecting variables for a final network model.
}
\examples{
1 + 1
}
