% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnetPowerSim.R
\name{sampleSize}
\alias{sampleSize}
\title{Tells the minimum possible sample size given some fixed p, m, and lags}
\usage{
sampleSize(p, m = 0, lags = 0, print = TRUE)
}
\arguments{
\item{p}{numeric}

\item{m}{numeric}

\item{lags}{numeric}

\item{print}{logical}
}
\value{
sample size minimum
}
\description{
Good for stuff
}
\examples{
1 + 1
}
