% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{Shows which variables were selected for each node of the network}
\usage{
selected(object, threshold = FALSE, mod = 1)
}
\arguments{
\item{object}{fitNetwork object}

\item{threshold}{numeric or logical}

\item{mod}{not sure}
}
\value{
A table
}
\description{
Shows which variables were selected for each node of the network
}
\examples{
1 + 1
}
