% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNet.R
\name{simNet}
\alias{simNet}
\title{Simulate network structure and data}
\usage{
simNet(
  N = 100,
  p = 5,
  m = FALSE,
  m2 = 0.1,
  b1 = NULL,
  b2 = NULL,
  sparsity = 0.5,
  intercepts = NULL,
  nIter = 250,
  msym = FALSE,
  onlyDat = FALSE,
  pbar = TRUE,
  div = 10,
  gibbs = TRUE,
  ordinal = FALSE,
  nLevels = 5,
  mord = FALSE,
  time = TRUE,
  mbinary = FALSE,
  minOrd = 3,
  m1 = NULL,
  m1_range = NULL,
  m2_range = c(0.1, 0.3),
  modType = "none",
  lags = NULL,
  V = 2,
  skewErr = FALSE,
  onlyNets = FALSE,
  netArgs = NULL,
  nCores = 1,
  cluster = "SOCK",
  getChains = FALSE,
  const = 1.5,
  fixedPar = NULL,
  V2 = 1,
  ...
)
}
\arguments{
\item{N}{Sample size}

\item{p}{Number of nodes}

\item{m}{moderator}

\item{m2}{numeric}

\item{b1}{vector or matrix?}

\item{b2}{vector or matrix?}

\item{sparsity}{numeric}

\item{intercepts}{vector}

\item{nIter}{numeric}

\item{msym}{logical}

\item{onlyDat}{logical}

\item{pbar}{logical}

\item{div}{numeric}

\item{gibbs}{logical}

\item{ordinal}{logical}

\item{nLevels}{numeric}

\item{mord}{logical}

\item{time}{logical}

\item{mbinary}{logical}

\item{minOrd}{numeric}

\item{m1}{numeric}

\item{m1_range}{numeric}

\item{m2_range}{numeric}

\item{modType}{character}

\item{lags}{numeric or logical}

\item{V}{numeric}

\item{skewErr}{logical or numeric}

\item{onlyNets}{logical}

\item{netArgs}{list}

\item{nCores}{numeric}

\item{cluster}{character}

\item{getChains}{logical}

\item{const}{numeric}

\item{fixedPar}{something}

\item{V2}{numeric}

\item{...}{other arguments}
}
\value{
A buncha stuff
}
\description{
Has some problems
}
\examples{
1 + 1
}
