% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNet.R
\name{simNet2}
\alias{simNet2}
\title{A wrapper for simNet to prevent failure}
\usage{
simNet2(
  ...,
  nets = NULL,
  maxiter = 10,
  pbar = TRUE,
  div = 10,
  divup = TRUE,
  maxdiv = 10^3,
  errors = FALSE,
  nostop = FALSE
)
}
\arguments{
\item{...}{simNet arguments}

\item{nets}{something}

\item{maxiter}{numeric}

\item{pbar}{logical}

\item{div}{numeric}

\item{divup}{logical}

\item{maxdiv}{numeric}

\item{errors}{logical}

\item{nostop}{logical}
}
\value{
simNet output
}
\description{
Possibly not even prevent, just stall.
}
\examples{
1 + 1
}
