% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNet.R
\name{summary.bootNet}
\alias{summary.bootNet}
\title{Descriptive statistics for bootNet}
\usage{
\method{summary}{bootNet}(object, centrality = TRUE, ...)
}
\arguments{
\item{object}{bootNet object}

\item{centrality}{logical}

\item{...}{other arguments}
}
\value{
A table of descriptives
}
\description{
Description
}
\examples{
1 + 1
}
