% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnetPowerSim.R
\name{summary.mnetPower}
\alias{summary.mnetPower}
\title{Descriptive statistics for power simulation results}
\usage{
\method{summary}{mnetPower}(object, ind = "all", order = NULL, decreasing = FALSE, ...)
}
\arguments{
\item{object}{mnetPowerSim output}

\item{ind}{performance indices}

\item{order}{character}

\item{decreasing}{logical}

\item{...}{Additional arguments.}
}
\value{
summary table
}
\description{
Description
}
\examples{
1 + 1
}
