% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURfunctions.R
\name{SURll}
\alias{SURll}
\title{Log-likelihood of SUR model with LRT to compare models}
\usage{
SURll(
  net0,
  net1 = NULL,
  nodes = FALSE,
  lrt = NULL,
  all = FALSE,
  d = 4,
  alpha = 0.05,
  s = "res",
  orderBy = NULL,
  decreasing = TRUE,
  sysfits = FALSE
)
}
\arguments{
\item{net0}{network}

\item{net1}{network}

\item{nodes}{logical}

\item{lrt}{logical}

\item{all}{logical}

\item{d}{numeric}

\item{alpha}{numeric}

\item{s}{character}

\item{orderBy}{character}

\item{decreasing}{logical}

\item{sysfits}{logical}
}
\value{
Results about SUR network
}
\description{
Similar to modLL
}
\examples{
1 + 1
}
