% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURfunctions.R
\name{SURtable}
\alias{SURtable}
\title{Obtain all possible LRTs comparing a list of SUR models}
\usage{
SURtable(
  fits,
  nodes = FALSE,
  orderBy = TRUE,
  d = 4,
  alpha = 0.05,
  decreasing = TRUE,
  names = NULL,
  rmsea = FALSE,
  s = "res"
)
}
\arguments{
\item{fits}{list of networks}

\item{nodes}{logical}

\item{orderBy}{logical or character?}

\item{d}{numeric}

\item{alpha}{numeric}

\item{decreasing}{logical}

\item{names}{logical or character?}

\item{rmsea}{logical}

\item{s}{character}
}
\value{
Table of LRTs
}
\description{
Includes RMSEAs
}
\examples{
1 + 1
}
