% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelect.R
\name{varSelect}
\alias{varSelect}
\title{Variable selection applied to network edges}
\usage{
varSelect(
  data,
  m = NULL,
  criterion = "AIC",
  method = "glmnet",
  lags = NULL,
  exogenous = TRUE,
  type = "g",
  center = TRUE,
  scale = FALSE,
  gamma = 0.5,
  nfolds = 10,
  varSeed = NULL,
  useSE = TRUE,
  nlam = NULL,
  covs = NULL,
  verbose = TRUE,
  beepno = NULL,
  dayno = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{m}{numeric}

\item{criterion}{character}

\item{method}{character}

\item{lags}{numeric or logical}

\item{exogenous}{logical}

\item{type}{character}

\item{center}{logical}

\item{scale}{logical}

\item{gamma}{numeric}

\item{nfolds}{numeric}

\item{varSeed}{numeric}

\item{useSE}{logical}

\item{nlam}{numeric}

\item{covs}{logical?}

\item{verbose}{logical}

\item{beepno}{something}

\item{dayno}{something}
}
\value{
Output from procedure
}
\description{
Returns output from a given variable selection procedure
}
\examples{
1 + 1
}
