% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetwork.R
\name{getFitCIs}
\alias{getFitCIs}
\title{Provides model coefficients with CIs}
\usage{
getFitCIs(fit, allNames = NULL, alpha = 0.05)
}
\arguments{
\item{fit}{fitNetwork object}

\item{allNames}{not sure}

\item{alpha}{error level}
}
\value{
List of tables
}
\description{
Requires that either 'fitobj' or 'SURfit' is included in the object
}
\examples{
1 + 1
}
