% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modnets.R
\docType{package}
\name{modnets}
\alias{modnets}
\title{modnets: Modeling Moderated Networks}
\description{
The modnets package is all about fitting and analyzing moderated network
models. Currently in development -- more updates and details to come.
}
\section{Details}{


\tabular{ll}{
Package: \tab modnets \cr
Title: \tab Modeling Moderated Networks \cr
Version: \tab 0.9.0 \cr
Author: \tab Trevor Swanson \cr
Maintainer: \tab \href{mailto:trevorswanson222@gmail.com}{trevorswanson222@gmail.com} \cr
URL: \tab \url{https://github.com/tswanson222/modnets} \cr
BugReports: \tab \url{https://github.com/tswanson222/modnets/issues} \cr
License: \tab GPL-3 \cr
Imports: \tab abind, corpcor, ggplot2, glinternet, glmnet, gridExtra, gtools, igraph, interactionTest, leaps, lme4, lmerTest, Matrix, methods, mvtnorm, parallel, pbapply, plyr, psych, qgraph, reshape2, systemfit \cr
Suggests: \tab sn, arm, hierNet, psychTools \cr
Date: \tab 2021-09-26
}
}

\references{
Swanson, T. J. (2020). \emph{Modeling moderators in psychological networks} (Publication No. 28000912) [Doctoral dissertation, University of Kansas]. ProQuest Dissertations Publishing.
}
\author{
Trevor Swanson

\strong{Maintainer:} Trevor Swanson \href{mailto:trevorswanson222@gmail.com}{trevorswanson222@gmail.com}
}
