% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plot.resample}
\alias{plot.resample}
\title{Plot method for output of resample function}
\usage{
\method{plot}{resample}(x, what = "network", ...)
}
\arguments{
\item{x}{output from resample function}

\item{what}{network, bootstrap, or coefs}

\item{...}{other arguments}
}
\value{
A plot
}
\description{
Description
}
\examples{
1 + 1
}
