% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotNet}
\alias{plotNet}
\title{Plot moderated network models}
\usage{
plotNet(
  object,
  which.net = "temporal",
  threshold = FALSE,
  layout = "spring",
  predict = FALSE,
  mnet = FALSE,
  names = TRUE,
  nodewise = FALSE,
  scale = FALSE,
  lag = NULL,
  con = "R2",
  cat = "nCC",
  plot = TRUE,
  elabs = FALSE,
  elsize = 1,
  rule = "OR",
  binarize = FALSE,
  mlty = TRUE,
  mselect = NULL,
  ...
)
}
\arguments{
\item{object}{network}

\item{which.net}{character}

\item{threshold}{numeric or logical}

\item{layout}{character}

\item{predict}{logical or network}

\item{mnet}{logical}

\item{names}{logical}

\item{nodewise}{logical}

\item{scale}{logical}

\item{lag}{numeric or logical}

\item{con}{character}

\item{cat}{character}

\item{plot}{logical}

\item{elabs}{logical}

\item{elsize}{numeric}

\item{rule}{character}

\item{binarize}{logical}

\item{mlty}{logical}

\item{mselect}{something}

\item{...}{other arguments}
}
\value{
A plot
}
\description{
Core function for plotting various types of network models.
}
\examples{
1 + 1
}
