% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNet.R
\name{summary.bootNet}
\alias{summary.bootNet}
\title{Descriptive statistics for \code{bootNet} objects}
\usage{
\method{summary}{bootNet}(object, centrality = TRUE, ...)
}
\arguments{
\item{object}{\code{bootNet} object}

\item{centrality}{Logical. Determines whether or not strength centrality and
expected influence should be computed for output.}

\item{...}{Additional arguments.}
}
\value{
A table of descriptives for bootNet objects.
}
\description{
Currently only works for unmoderated GGMs.
}
\examples{
\dontrun{
x <- bootNet(data)
summary(x)
}
}
