% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURfunctions.R
\name{SURfit}
\alias{SURfit}
\title{Fit SUR model with or without constraints}
\usage{
SURfit(
  data,
  varMods = NULL,
  mod = "min",
  maxiter = 100,
  m = NULL,
  type = "g",
  center = TRUE,
  scale = FALSE,
  exogenous = TRUE,
  covs = NULL,
  sur = TRUE,
  consec = NULL,
  ...
)
}
\arguments{
\item{data}{dataset}

\item{varMods}{output of varSelect}

\item{mod}{character}

\item{maxiter}{numeric}

\item{m}{moderator}

\item{type}{character}

\item{center}{logical}

\item{scale}{logical}

\item{exogenous}{logical}

\item{covs}{something}

\item{sur}{logical}

\item{consec}{something}

\item{...}{other arguments}
}
\value{
Models
}
\description{
Actually not sure if this is a top-level function.
}
\examples{
1 + 1
}
