% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURfunctions.R
\name{SURnet}
\alias{SURnet}
\title{Creates temporal and contemporaneous network of SUR results}
\usage{
SURnet(
  fit,
  dat,
  s = "sigma",
  m = NULL,
  threshold = FALSE,
  mval = NULL,
  medges = 1,
  pcor = "none"
)
}
\arguments{
\item{fit}{network}

\item{dat}{data}

\item{s}{character}

\item{m}{numeric}

\item{threshold}{logical}

\item{mval}{numeric}

\item{medges}{numeric}

\item{pcor}{character}
}
\value{
Temporal and contemporaneous networks
}
\description{
Used after something
}
\examples{
1 + 1
}
